requirements.txt
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
db-dtypes==1.0.3
Fiona==1.8.21
Flask==2.2.2
geopandas==0.11.1
google-api-core==2.8.2
google-auth==2.11.0
google-cloud-bigquery==3.3.2
google-cloud-bigquery-storage==2.14.2
google-cloud-core==2.3.2
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
greenlet==1.1.3
grpcio==1.47.0
grpcio-status==1.47.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
munch==2.5.0
numpy==1.23.2
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
playsound==1.3.0
proto-plus==1.22.0
protobuf==4.21.5
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydub==0.25.1
PyMySQL==1.0.2
pyparsing==3.0.9
PyPDF2==2.10.4
pyproj==3.3.1
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
rsa==4.9
Shapely==1.8.4
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.40
urllib3==1.26.12
Werkzeug==2.2.2
xlrd==2.0.1
